refactor(board): isolate feature tests from the app shell
Feature tests need the real Language/Theme/Snackbar stack but can't import
@app. Hoist the composition to @shared/providers/app-providers so main.tsx
and the board tests share one definition instead of a hand-synced
@shared/testing copy, and collapse the feature import-isolation eslint rule
into a single block covering source and tests.
