+++ src/features/board/board-viewer.test.tsx
@@ -1,7 +1,7 @@
-import { AppProviders } from "@app/app-providers";
 import { expectNoA11yViolations } from "@shared/testing/axe";
 import { TEST_IMAGE_SRC } from "@shared/testing/fixtures";
 import { stubAudio, stubSpeech } from "@shared/testing/device-output";
+import { AppProviders } from "@shared/providers/app-providers";
 import type { OBFBoard } from "@shayc/open-board-format";
 import { MemoryRouter } from "react-router";
 import { beforeEach, describe, expect, test, vi } from "vitest";
